# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bookparking — API",
    "version": "1.0.0",
    "description": "API de bookparking : comptes, sessions, annonces de places de parking et demandes de location.\n\nToutes les erreurs sont rendues par le filtre d'exception NestJS sous la forme `{ \"statusCode\": number, \"message\": string }`. Les messages sont en français et proviennent soit du décodage `effect/Schema` (400), soit des erreurs du domaine.\n\nL'authentification se fait par un jeton d'accès porté en `Authorization: Bearer <token>`, obtenu via `POST /session`."
  },
  "servers": [
    {
      "url": "http://localhost:{port}",
      "description": "Environnement local",
      "variables": {
        "port": {
          "default": "3000",
          "description": "Variable d'environnement PORT (3000 par défaut)."
        }
      }
    }
  ],
...
  "paths": {
...
    "/rental-request": {
      "post": {
        "tags": [
          "rental-request"
        ],
        "operationId": "requestRental",
        "summary": "Demander une location",
        "description": "Demande la location d'une place — identifiée par le couple (adresse, box) — sur une période de jours calendaires, au nom du compte porté par le jeton, et ouvre la page de paiement Stripe au prix que l'api a figé.\n\nLa demande reste en attente de paiement jusqu'à ce que Stripe annonce l'empreinte ; le délai de `RENTAL_REQUEST_EXPIRY_IN_HOURS` (48 heures par défaut) court ensuite depuis l'empreinte. Une page de paiement vaut trente minutes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestRentalRequest"
              },
              "examples": {
                "nominal": {
                  "value": {
                    "address": "12 rue des Lilas, 75011 Paris",
                    "box": "B12",
                    "fromDay": "2026-10-01",
                    "toDay": "2026-10-03"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Demande enregistrée, en attente de paiement.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestRentalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Corps de requête invalide, ou en-tête `Idempotency-Key` absent ou qui n'est pas un UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "dateInvalide": {
                    "value": {
                      "statusCode": 400,
                      "message": "fromDay: Date de début invalide"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "description": "La demande est bien formée mais le domaine la refuse. Messages possibles : « Les dates demandées sont invalides », « La période demandée dépasse la durée maximale de 366 jours », « Ces dates sont déjà louées », « Cette place n'a aucune annonce publiée », « Aucun tarif ne couvre la période demandée ». « Cet identifiant de demande a déjà servi pour une autre place ou une autre période ».",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "datesDejaLouees": {
                    "value": {
                      "statusCode": 422,
                      "message": "Ces dates sont déjà louées"
                    }
                  },
                  "aucuneAnnoncePubliee": {
                    "value": {
                      "statusCode": 422,
                      "message": "Cette place n'a aucune annonce publiée"
                    }
                  },
                  "aucunTarif": {
                    "value": {
                      "statusCode": 422,
                      "message": "Aucun tarif ne couvre la période demandée"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "description": "Stripe ne répond pas : la demande est abandonnée et ses dates rendues.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Une demande sous cet identifiant d'intention est en cours de création : réessayer dans un instant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "L'identifiant d'intention, un UUID choisi par le client pour une place et une période. Renvoyé à l'identique, il rend la demande déjà créée — même identifiant, même page de paiement — au lieu d'en créer une seconde, et la réponse porte alors `Idempotent-Replayed: true`. Il est propre au compte qui l'envoie.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      },
      "get": {
        "tags": [
          "rental-request"
        ],
        "operationId": "listMyRentalRequests",
        "summary": "Lister les demandes que l'on a faites",
        "description": "Rend les demandes de location faites par le compte appelant, en tant que locataire.",
        "responses": {
          "200": {
            "description": "La liste, éventuellement vide.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RentalRequest"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/rental-request/received": {
      "get": {
        "tags": [
          "rental-request"
        ],
        "operationId": "listReceivedRentalRequests",
        "summary": "Lister les demandes reçues sur ses places",
        "description": "Rend les demandes faites sur les annonces du compte appelant, en tant que propriétaire. C'est **la seule route qui rend l'identifiant d'une demande** à un client, et donc la seule qui rende `POST /rental-request/{id}/confirmation` atteignable.",
        "responses": {
          "200": {
            "description": "La liste, éventuellement vide.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RentalRequest"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/rental-request/{id}/confirmation": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Identifiant de la demande de location (UUID).",
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "example": "9b2f4d6a-1c3e-4f5a-8b7c-0d1e2f3a4b5c"
        }
      ],
      "post": {
        "tags": [
          "rental-request"
        ],
        "operationId": "confirmRentalRequest",
        "summary": "Confirmer une demande de location",
        "description": "Confirme une demande de location. Seul le propriétaire de la place peut confirmer.\n\nUn identifiant mal formé, une demande inconnue et une demande appartenant à quelqu'un d'autre répondent toutes `404`, avec le même message.\n\nConfirmer prélève l'empreinte du conducteur. Si sa banque refuse le prélèvement, la confirmation répond `409` et la demande passe en `PAYMENT_FAILED`.",
        "responses": {
          "204": {
            "description": "Demande confirmée. Aucun corps de réponse."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Demande inexistante, identifiant mal formé, ou demande appartenant à un autre compte.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "introuvable": {
                    "value": {
                      "statusCode": 404,
                      "message": "Cette demande de location n'existe pas"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "La demande a expiré et ne peut plus être confirmée.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "expiree": {
                    "value": {
                      "statusCode": 409,
                      "message": "Cette demande de location a expiré et ne peut plus être confirmée"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "description": "Stripe ne répond pas : rien n'est prélevé ni confirmé, la confirmation peut être rejouée.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/rental-requests": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "listAdminRentalRequests",
        "summary": "Lister toutes les demandes de location",
        "description": "Tous statuts confondus, avec le propriétaire et le locataire.",
        "responses": {
          "200": {
            "description": "La liste.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminRentalRequest"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/admin/rental-requests/{id}/cancellation": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "adminCancelRentalRequest",
        "summary": "Annuler une demande de location",
        "description": "Confirmée ou non. **Aucun remboursement n'est émis** : le produit ne sait pas encaisser, donc il ne sait pas rendre. La place redevient libre — la contrainte d'exclusion ignore les demandes annulées.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerationRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Fait. L'action est inscrite au journal de modération."
          },
          "400": {
            "description": "Motif absent ou trop court.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Cible inexistante, ou déjà dans l'état demandé.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/rental-request/{id}/abandonment": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Identifiant de la demande de location (UUID).",
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "example": "9b2f4d6a-1c3e-4f5a-8b7c-0d1e2f3a4b5c"
        }
      ],
      "post": {
        "tags": [
          "rental-request"
        ],
        "operationId": "abandonRentalRequest",
        "summary": "Abandonner sa demande avant de payer",
        "description": "Appelée quand le conducteur revient de Stripe sans payer : la page de paiement est fermée et les dates rendues aussitôt. Seul l'auteur de la demande peut l'abandonner.",
        "responses": {
          "204": {
            "description": "Demande abandonnée, ou déjà abandonnée."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Demande inexistante, identifiant mal formé, ou demande d'un autre compte.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "L'empreinte est déjà posée : la demande attend le propriétaire.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/rental-request/{id}/cancellation": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Identifiant de la demande de location (UUID).",
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "example": "9b2f4d6a-1c3e-4f5a-8b7c-0d1e2f3a4b5c"
        }
      ],
      "post": {
        "tags": [
          "rental-request"
        ],
        "operationId": "cancelRental",
        "summary": "Annuler une réservation",
        "description": "Annule une demande qui attend le propriétaire ou une location confirmée, tant qu'elle n'a pas commencé. Le conducteur est remboursé en totalité jusqu'à l'échéance `freeCancellationUntil` incluse, et plus du tout après ; le propriétaire peut annuler à tout moment avant le début, et le conducteur récupère alors tout. Une empreinte non prélevée est toujours levée. Annuler deux fois rend le même effet sans rien refaire.",
        "responses": {
          "200": {
            "description": "Réservation annulée.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancellationResult"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Demande inexistante, identifiant mal formé, ou demande que ce compte ne peut ni voir ni annuler.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "La location a commencé, ou la demande n'est ni en attente du propriétaire ni confirmée.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Jeton rendu par `POST /session`, porté en `Authorization: Bearer <token>`. Un en-tête absent, mal préfixé, vide ou invalide répond `401`."
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Jeton d'accès absent, mal formé ou invalide.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "nonAutorise": {
                "value": {
                  "statusCode": 401,
                  "message": "Unauthorized"
                }
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Erreur inattendue.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "inattendue": {
                "value": {
                  "statusCode": 500,
                  "message": "Une erreur inattendue est survenue"
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Corps d'erreur rendu par le filtre d'exception NestJS.",
        "required": [
          "statusCode",
          "message"
        ],
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "Message en français. Pour une erreur de décodage (400), les échecs sont formatés `chemin: message`.",
            "example": "Annonce introuvable"
          }
        }
      },
...
      "RequestRentalRequest": {
        "type": "object",
        "required": [
          "address",
          "box",
          "fromDay",
          "toDay"
        ],
        "additionalProperties": false,
        "properties": {
          "address": {
            "type": "string",
            "example": "12 rue des Lilas, 75011 Paris"
          },
          "box": {
            "type": "string",
            "example": "B12"
          },
          "fromDay": {
            "type": "string",
            "format": "date",
            "description": "Premier jour loué, en jour calendaire `AAAA-MM-JJ`.",
            "example": "2026-10-01"
          },
          "toDay": {
            "type": "string",
            "format": "date",
            "description": "Dernier jour loué, en jour calendaire `AAAA-MM-JJ`. La période ne peut pas dépasser 366 jours.",
            "example": "2026-10-03"
          }
        }
      },
...
      "RentalRequest": {
        "type": "object",
        "description": "Une demande de location. Ni le locataire ni le propriétaire n'y figurent : les deux routes qui la rendent sont déjà clés sur le compte appelant.",
        "required": [
          "id",
          "listingId",
          "address",
          "box",
          "fromDay",
          "toDay",
          "priceInCents",
          "status",
          "money",
          "requestedAt",
          "confirmedAt",
          "startsAt",
          "freeCancellationUntil"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "L'identifiant à passer à `POST /rental-request/{id}/confirmation`."
          },
          "listingId": {
            "type": "string",
            "format": "uuid"
          },
          "address": {
            "type": "string"
          },
          "box": {
            "type": "string"
          },
          "fromDay": {
            "type": "string",
            "format": "date"
          },
          "toDay": {
            "type": "string",
            "format": "date"
          },
          "priceInCents": {
            "type": "integer",
            "description": "Le montant figé au moment de la demande."
          },
          "status": {
            "type": "string",
            "enum": [
              "AWAITING_PAYMENT",
              "PENDING",
              "CONFIRMED",
              "EXPIRED",
              "CANCELLED",
              "ABANDONED",
              "PAYMENT_FAILED"
            ],
            "description": "`AWAITING_PAYMENT` : la page de paiement est ouverte, l'empreinte n'est pas constatée ; la demande retient ses dates mais le propriétaire ne la voit pas. `ABANDONED` : le paiement n'a jamais eu lieu. `PAYMENT_FAILED` : la banque a refusé le prélèvement à la confirmation."
          },
          "money": {
            "type": "string",
            "enum": [
              "NONE",
              "AUTHORIZED",
              "CAPTURED",
              "RELEASE_DUE",
              "RELEASED",
              "REFUND_DUE",
              "REFUNDED"
            ],
            "description": "Où en est l'argent du conducteur. `AUTHORIZED` : empreinte posée, rien de prélevé. `CAPTURED` : prélevé à la confirmation. `RELEASE_DUE` / `REFUND_DUE` : l'empreinte est en cours de levée, le remboursement en cours. `NONE` : demande faite avant l'encaissement."
          },
          "requestedAt": {
            "type": "string",
            "format": "date-time"
          },
          "confirmedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "startsAt": {
            "type": "string",
            "format": "date-time",
            "description": "Le premier instant de la location, heure de Paris. Une réservation ne s'annule plus à partir de cet instant."
          },
          "freeCancellationUntil": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "L'échéance d'annulation gratuite, figée à la demande : jusqu'à cet instant inclus, le conducteur qui annule une location confirmée est remboursé en totalité."
          }
        }
      },
...
      "RequestRentalResponse": {
        "type": "object",
        "required": [
          "id",
          "checkoutUrl"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "L'identifiant de la demande, en attente de paiement."
          },
          "checkoutUrl": {
            "type": "string",
            "format": "uri",
            "description": "La page de paiement Stripe Checkout où le conducteur pose son empreinte. Toujours sur `https://checkout.stripe.com/`."
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}